Allow setting position increment one for subtoken.
This change allows to set the position increment of subtoken to one.

For AND searches including the subtoken this is required, while they
are not considered otherwise.

The option is set to zero (current behaviour) by default while it can be
turned on setting subtokenPositionIncrement to one.
Turning this on will not return the original token while the subtokens
will be returned with an index of one. In a AND query searching for
orangenschokolade and a dict with orangen and schokolade this would
result in a query for orangen AND schokolade.
With the current behaviour this would not be possible since the subtoken
orangen and schokolade are not considered in the AND query.